requirements.txt
altgraph==0.17
certifi==2020.6.20
cycler==0.10.0
future==0.18.2
kiwisolver==1.2.0
matplotlib==3.2.2
minimalmodbus==1.0.2
numpy==1.19.1
pefile==2019.4.18
Pillow==7.2.0
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.7
pyparsing==2.4.7
Pypubsub==4.0.3
pyserial==3.4
PySide2==5.15.0
python-dateutil==2.8.1
pywin32-ctypes==0.2.0
shiboken2==5.15.0
six==1.15.0
